big update -- see description
- added gyroscope tracking functionality
- added file creation capability to save gyroscope data
- added 15 second timer
- tracks user input for 3 seconds every 15 seconds (per timer)

basically:
assuming the gyroscope is available in the iPhone, the app tracks the gyroscope for 3 seconds every 15 seconds and prompts the user to save that data into a csv file that they can select on their computer.

in future update, data should be automatically saved into computer, not user-prompted. for now, it's like this b/c bypassing iTunes.